frr bgp conf template handling switch_type == voq
when switch_type is voq, it could either be a single asic or multi-asic
or even a voq_chassis
bgpd.main.conf.j2 needs to handle the single asic case differently than
the multi-asic or voq_chassis.
A new filter file_exists is now used within j2 files to look for
chasssisdb.conf and determine if its a voq_chassis
Incase of unit test, the test just runs with a json file only. Hence
using a chassisdb_conf_present in it to represent the chassis case

When assigning the loopback0 address, handle the single asic voq case.