Route Policy PREFIX list implementation issues related to `sequence_number` (#22888)
Why I did it
The sequence_number is used as a key in the yang model and is therefore mandatory, however the implementation does not expect a sequence id at all and just uses the order in which the prefixes appear.

The sequence_number however may be a desirable feature so should be honored if we have it, but for compatibility with any existing users (who clearly are not using YANG validation), it must be removed from the yang list model as a key.

This also being a key also does not align with the HLD: https://github.com/sonic-net/SONiC/blob/master/doc/mgmt/SONiC_Design_Doc_Unified_FRR_Mgmt_Interface.md#321121-prefix

If passing value as YANG requires, the jinja2 renderer and frrcfgd both fail.

This affects unified mode with frr_mgmt_framework_config.

How did I do it?

This updates the YANG model to remove the sequence_number as a key, adds a uniqueness must clause validation for the sequence_number. Tests are also implemented to confirm behavior. frr-mgmt-framework has been expanded to include support for the sequence_number as an optional parameter.

Signed-off-by: Feng Pan <fenpan@microsoft.com>