PDDF: Do not always use -f to load custom PDDF modules (#23732)
When in secureboot, -f will fail with a module signing issue even
though the module is signed properly.  This implies it is doing
some module modification to force it to be loaded which invalidates
the signing.

First try to load custom modules without the -f flag, then on failure
use the -f flag to maintain compatibility with the existing
implementation.

Signed-off-by: Feng Pan <fenpan@microsoft.com>