PDDF: add pddfparse flag to create or delete subtree (#23684)
We need an easy way to create or delete just a subset of the entire PDDF
setup for a particular system.

This can be useful when you want to upgrade a component that is
currently providing some functionality through PDDF. In this way, we can
delete the subtree provided by the component, upgrade it (which causes
it to be unresponsive for a while), then re-create the bits of PDDF that
were using this component.

Co-authored-by: lotus-nexthop <lotus@nexthop.ai>
Signed-off-by: Feng Pan <fenpan@microsoft.com>