libyang3: high frequency telemetry test fix
Its not entirely clear what this test was attempting to
validate.  Its possible they wanted to ensure the lists
weren't empty, but the yang model itself does not have that
restriction.  It would be out of scope for this change to
modify an invalid yang model.

The test was passing a string to a list which is invalid,
so just modified the test to validate both object_names and
object_counters will error out if a string is provided.

Signed-off-by: xiaweijiang <xiaweijiang@microsoft.com>