find_data_dependencies() should be recursive.
The only caller to find_data_dependencies() is in sonic-utilities
and it is called in a recursive manner by calling into libyang
functions directly.  This should be properly abstracted within
sonic-yang-mgmt to ensure it is not dependent on any particular
libyang version.

Test cases have been updated to reflect the behavior while still
maintaining compatibility with the existing sonic-utilities until
it is updated.