Fix cargo-tarpaulin install in sonic-slave-bookworm
Package ruzstd is a transitive dependency of cargo-tarpaulin. ruzstd v0.8.2 was
released on 2025-11-02, and it uses new api `is_multiple_of` that has been
stabilised and released in rust 1.87.0. This method is marked as unstable in
older rust versions, and hence fails the build of sonic-slave-bookworm.

Bump rust version to 1.87.0 to address this issue.

Signed-off-by: Krzysztof Kościuszkiewicz <k.kosciuszkiewicz@arista.com>