Add build flag to skip build checks
Part of the standard process for building a debian package is to run
`make check` which usually runs some self tests of the package being
built. We already know to skip this when cross compiling, add an
additional SONIC_BUILD_NOCHECK that allows the `make check` step to be
skipped even when building natively.

There are merits to having the checks run by default and this does not
change any behaviour in that respect. It does however provide developers
with an escape hatch to speed up local builds.

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>