[Arista] Reset switch chip in Aboot for QuicksilverP512 and QuicksilverP (#23939)
Why I did it
Mitigate for the recent issue that under some unknown initial conditions, the 7060X6-64PE-B device could get into reboot loop. It's been identified so far that the reboot loop is related to the some bad switch chip state. While still trying to figure out the initial trigger, we propose this mitigation that break the reboot loop. The same mitigation to 7060X6-64PE is also included in this change.

Work item tracking
Microsoft ADO (number only):
How I did it
In boot0 script, if the environment is Aboot and the platform is 7060X6-64PE-B, reset the switch chip.

How to verify it
According to our test, this doesn't affect normal booting.

Signed-off-by: xiaweijiang <xiaweijiang@microsoft.com>