Dockerfile.j2 slave should include optional vendor-specific file
There are instances where a vendor may want to add additional
packages to the slave docker that aren't suitable for upstreaming
such as necessary applications or packages for secureboot signing
unique to the vendor environment.

This allows vendors to define their own `sonic-slave-Dockerfile.vendor.j2`
file within the toplevel files folder if they want to use this
feature.

Since these are things that would never be upstreamed, it eases
the maintenance burden of handling merge conflicts as well as
should be automatically updated when new slave containers are
defined (e.g. bookworm -> trixie).

This follows the precident where a vendor may define their own
config settings in rules/config.user.

Signed-off-by: Brad House <bhouse@nexthop.ai>