[dhcprelay] Don't build flaky unit tests
The dhcp6relay unit tests can trigger an endless loop of failures in
some build environments:

  make[2]: Leaving directory '/sonic/src/dhcprelay'
     dh_auto_test
      make -j20 test
  make[2]: Entering directory '/sonic/src/dhcprelay'
  sudo ASAN_OPTIONS=detect_leaks=0 ./build-test/dhcp6relay-test --gtest_output=xml:build-test/dhcp6relay-test-test-result.xml || true
  gcovr -r ./ --html --html-details -o build-test/dhcp6relay-test-code-coverage.html
  gcovr -r ./ --xml-pretty -o build-test/dhcp6relay-test-code-coverage.xml
  AddressSanitizer:DEADLYSIGNAL
  AddressSanitizer:DEADLYSIGNAL
  ...

Technically this is an AddressSanitizer bug. It's been seen with GCC
11.4.0 and GCC 13.2.0 but other versions (and clang) may be affected. It
is possible to run the unit tests on affected systems by setting the
number of ASLR bits to 28 (e.g. `sudo sysctl -w vm.mmap_rnd_bits=28`)
but doing so requires being able to adjust this setting in the first
place and would mean the SONiC build process affects the build machine
which is not desirable. The best thing sonic-build-image can do is to
disable the unit tests for the dhcprelay package.

Fixes #19007
Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>