[build]: Fix cargo-tarpaulin install in sonic-slave-bookworm (#24428)
Package ruzstd is a transitive dependency of cargo-tarpaulin. ruzstd v0.8.2 was
released on 2025-11-02, and it uses new api `is_multiple_of` that has been
stabilised and released in rust 1.87.0. This method is marked as unstable in
older rust versions, and hence fails the build of sonic-slave-bookworm.

Use locked dependency version when installing cargo-tarpaulin to avoid using
ruzstd v0.8.1, as sonic-slave-bookworm uses rust 1.86.0.

Signed-off-by: Krzysztof Kościuszkiewicz <k.kosciuszkiewicz@arista.com>
Signed-off-by: Feng Pan <fenpan@microsoft.com>