Fix setting the hostname from a DHCPv6 response
The FQDN in a DHCPv6 response is sent via option 39, which is encoded to
include some additional information about who can update the hostname of
devices. This means that the `dhcp6_fqdn` option used here is just a
string of hex-encoded bytes, not the actual hostname. The actual decoded
hostname is accessible via `fqdn_fqdn`.

Update the `sethostname6` script to use `new_fqdn_fqdn` instead of
`neq_dhcp6_fqdn`.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>