Remove nodesource nodejs sources for Bookworm arm container
Debian Bookworm already has a nodejs that is newer than the
nodesource.com version 14 that the Dockerfile is currently installing.
Bookworm has nodejs 18.

Because of this, we run the script (curl | bash) from nodesource.com for
nodejs 14 which makes the process wait for 80 seconds because the script
is deprecated. Then when we actually install nodejs, we get the version
from Bookworm's repositories instead of nodesource.com anyway. This
makes the process of using nodesource.com wasteful and pointless.

The easy fix is to just remove this line where we add the nodesource
repo.

Signed-off-by: Nate White <nate@nexthop.ai>