[bgpcfgd]: Fix bgpcfgd uDT46 SID configuration for default VRF
There is a bug in bgpcfgd: it cannot configure a uDT46 SID for the
default VRF.

For example, if you add the following entry to the Config DB:

```
"SRV6_MY_SIDS": {
    "loc1|FCBB:BBBB:20:F1::": {
        "action": "uDT46",
        "vrf": "default",
        "decap_dscp_mode": "uniform"
    }
}
```

bgpcfgd generates this FRR configuration:

```
segment-routing
 srv6
  static-sids
   sid fcbb:bbbb:1:fe40::/64 locator loc1 behavior uDT46
```

However, FRR rejects this command and returns the following error:

```
% Unknown command sid fcbb:bbbb:1:fe40::/64 locator MAIN behavior uDT46
```

The root cause is that bgpcfgd omits the required `vrf` parameter in
the FRR configuration when the VRF is set to "default". FRR, however,
always expects the `vrf` parameter to be specified explicitly, even for
the default VRF.

The correct command should be:

```
sid fcbb:bbbb:1:fe40::/64 locator loc1 behavior uDT46 vrf default
```

This pull request addresses the problem by ensuring that bgpcfgd always
includes the `vrf` parameter in the FRR configuration, regardless of
whether the VRF is set to "default" or another value.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>