Remove outdated comment for Flask and unpin package six
Because Werkzeug 3.1.3 has a known issue, previously Flask and Werkzeug versions are pinned down as below:
Flask==3.0.3
Werkzeug==3.1.2

Now the latest versions of Flask and Werkzeug on bookworm is as below:
    Flask 3.1.2
    Werkzeug 3.1.4

The comment for Werkzeug 3.1.3 is no longer relevant. This commit removed the outdated comment.

The six package version is also un-pinned in this commit.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>