[FPM]: Fix dplane_fpm_sonic for multiple SRv6 nexthops
Currently, there is a bug in the dplane_fpm_sonic to handle ECMP routes. When FRR has an ECMP route with multiple SRv6 nexthops, the dplane_fpm_sonic encodes the route with only the first SRv6 nextop in the Netlink message, ignoring the other nexthops.

As a result, the Netlink message sent to fpmsyncd leads to the installation of the route using only the first SRv6 nexthop, breaking the EMCP nature of SRv6 routes.

This pull request fixes the issue in the dplane_fpm_sonic to handle all SRv6 nexthops in the route.