Build: Fix error printed out by Git when parsing .dep files
During a build or configure, there were 2 prints of this message:

fatal: not a git repository (or any of the parent directories): .git

This is confusing because during a build process if there is a fatal
error calling `git` this probably means the build is unsound.
Accordingly, people were confused about errors being printed even though
the rest of the build appeared to succeed.

In this case, I don't believe the build was unsound, unless one was
using the DPKG Caching Framework as Makefile.cache appears to be the
only place where these .dep files are read. I think in both cases here,
the DPKG Caching Framework would refuse to cache these packages, thus
the build would still be sound, but some caching would be missed.

The fix is to ensure that we call git ls-files with a real path to the
source for these packages. This way the path will be a valid git
repository.

Signed-off-by: Nate White <nate@nexthop.ai>