[bgpcfgd]: Fix missing SRv6 encapsulation source address
Currently, there is no way to configure the source address for SRv6
encapsulation in bgpcfgd. As a result, FRR defaults to using an
incorrect source address.

This incorrect source address is then pushed to SONiC fpmsyncd, leading
to the data plane using the wrong source address in encapsulated packets.

To resolve this issue, this PR extens bgpcfgd to allow users to configure
the SRv6 encapsulation source address.

With this enhancement, the correct source address will be programmed and
used throughout the control and data plane.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>