Ensure INCLUDE_STP is set and is passed through when building
INCLUDE_STP wasn't getting passed through from the top-level make.
Ensure that it gets passed through so it can be enabled or disabled as
required. The current not explicitly disabled or enabled state is
maintained. This is effectively the same as disabled as the checks all
use positive logic.

Fixes #24768
Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>