Platform support for Arista-720DT-48S-MGX (#24529)
Why I did it
To add support for the Arista-720DT-48S-MGX platform

How I did it
Added a device folder and relevant config files for the platform. Added the platform to boot0. Also expanded a workaround in files/scripts/syncd.sh to apply to all 720DT-48S models.

How to verify it
Switch boots. All ports are up and passing traffic.

Signed-off-by: Justin Oliver <justinoliver@arista.com>
Signed-off-by: xiaweijiang <xiaweijiang@microsoft.com>