Increase lldp netlink buffer limits (#24752)
Why I did it
When enough neighbors exit for a switch (>211), LLDP will stop reporting the last few neighbors after swss restarts. This happens even though the tcp packets from those neighbors are still arriving over the ports. This is because the netlink buffer is filled up when the restart occurs and the netlink delete messages are not received by lldp, this leads to the service not including the new neighbors over those ports, which will persist until lldp is restarted:

How I did it
By increasing both the minimum and maximum netlink receive buffer in LLDP I was able to fix this issue

How to verify it
Run:

show lldp table
systemctl restart swss

Check without the change to see if the lldp buffer cannot handle all of the neighbors, then install a version of Sonic with the change to check that the larger buffer fixes this issue.

Signed-off-by: Connor Roos <croos@nvidia.com>
Signed-off-by: Feng Pan <fenpan@microsoft.com>