Issue 25728: Add vrf ordering dependency to chrony config
When chrony is configured to start in the mgmt vrf, it needs to be
ordered after interfaces-config.service, since that's the service that
sets up the mgmt vrf.

NTP/chrony service is not coming up correctly after reboot/reload when
configured to start in the mgmt vrf.

When being run in the mgmt vrf, chrony needs to wait to start up until
the mgmt vrf is created, so add an `After` dependency in the chrony
config override file that waits for the appropriate service to finish
(interfaces-config.service).

Manually verified. Configure the device to enable the mgmt vrf and to
have ntp configured to use the mgmt vrf. Then issue a reboot or 'config
reload'. Without these changes, chrony will come up incorrectly and
require a restart to function (this is a race condition, so the exact
failure can be different, and it can even work occasionally, but fails
pretty consistently). With these changes, reboot/reload work as
expected.

Signed-off-by: Anders Linn <anders@nexthop.ai>